+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheetFormFactory/PaymentSheetFormFactory+Mandates.swift
@@ -127,6 +127,10 @@ extension PaymentSheetFormFactory {
         let mandateText = String(format: String.Localized.touch_n_go_mandate_text, configuration.merchantDisplayName)
         return makeMandate(mandateText: mandateText)
     }
+    func makeTrueMoneyMandate() -> SimpleMandateElement {
+        let mandateText = String(format: String.Localized.truemoney_mandate_text, configuration.merchantDisplayName)
+        return makeMandate(mandateText: mandateText)
+    }
     func makeKoreanPaymentMethodMandate() -> SimpleMandateElement {
         let mandateText = String(
             format: String.Localized.korean_payment_method_mandate_text,