+++ StripePayments/StripePayments/Source/API Bindings/Models/PaymentMethods/STPPaymentMethodEnums.swift
@@ -130,6 +130,8 @@ import Foundation
     case ngBankTransfer
     /// A Touch 'n Go payment method
     case touchNGo
+    /// A TrueMoney payment method
+    case trueMoney
     /// An unknown type.
     case unknown
 
@@ -256,6 +258,8 @@ import Foundation
             return "Naira bank transfer"
         case .touchNGo:
             return "Touch 'n Go"
+        case .trueMoney:
+            return "TrueMoney"
         case .goPay:
             return "GoPay"
         case .cardPresent,
@@ -387,6 +391,8 @@ import Foundation
             return "ng_bank_transfer"
         case .touchNGo:
             return "touch_n_go"
+        case .trueMoney:
+            return "truemoney"
         case .goPay:
             return "gopay"
         }