+++ StripePayments/StripePayments/Source/API Bindings/Models/PaymentMethods/Types/STPPaymentMethodTrueMoney.swift
@@ -1,0 +1,38 @@
+//
+//  STPPaymentMethodTrueMoney.swift
+//  StripePayments
+//
+//  Created by Nick Porter on 9/16/26.
+//  Copyright © 2026 Stripe, Inc. All rights reserved.
+//
+
+import Foundation
+@_spi(STP) import StripeCore
+
+/// A TrueMoney Payment Method.
+/// - seealso: https://docs.stripe.com/payments/truemoney/accept-a-payment?payment-ui=direct-api
+public class STPPaymentMethodTrueMoney: NSObject, STPAPIResponseDecodable {
+    /// :nodoc:
+    @objc private(set) public var allResponseFields: [AnyHashable: Any] = [:]
+
+    /// :nodoc:
+    @objc public override var description: String {
+        let props = [
+            String(format: "%@: %p", NSStringFromClass(STPPaymentMethodTrueMoney.self), self)
+        ]
+        return "<\(props.joined(separator: "; "))>"
+    }
+
+    /// :nodoc:
+    @objc public class func decodedObject(fromAPIResponse response: [AnyHashable: Any]?) -> Self? {
+        guard let response else {
+            return nil
+        }
+        return self.init(dictionary: response)
+    }
+
+    required init(dictionary: [AnyHashable: Any]) {
+        super.init()
+        allResponseFields = dictionary
+    }
+}